adminui/requirements.txt
aiohttp==3.10.11
aiohttp-cors==0.7.0
aiohttp-middlewares==2.4.0
aiohttp-jinja2==1.5
aiohttp-session==2.9.0
cryptography==41.0.5
PyJWT==2.7.0
python-dotenv==1.0.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
loguru==0.7.2
pyyaml==6.0.1
dataclasses-json==0.5.6
pymongo[srv]==4.0.1
pyopenssl==23.3.0
coverage==7.6.0
requests==2.25.1
StrEnum==0.4.7
pyhumps==3.8.0
paramiko==2.7.2
six==1.16.0
progressbar==2.5
tqdm==4.62.3
obonet==0.3.1
aiohttp-autoreload==0.0.1
django-allauth[idp-oidc]==65.11.0
django-browser-reload==1.16.0
Django==4.2.16
python-dotenv==1.0.1
requests==2.25.1
